# CRDB: detect draining nodes on connection acquisition