At higher scales, creating new connections against CRDB dominates SpiceDB latency. 

We currently follow best-practices around connection draining by ensuring that our connection max lifetimes are <5m, which ensures all connections are drained properly during node drains in Cockroach Dedicated.

As we add more SpiceDB nodes and CRDB nodes, we increase the number of connections to take advantage of the extra capacity. CRDB connection latency increases as connection rate increases.

This PR checks the connection when we acquire it from the pool. If the node is draining, the connection is closed and we grab a new one instead. This allows us to extend lifetimes past 5m and still handle node drains.

This has a negative performance impact on CRDB at low scales - detecting when a node is draining requires another roundtrip, and this comes through in the benchmarks:

```go
goos: darwin
goarch: arm64
pkg: github.com/authzed/spicedb/internal/datastore/benchmark
                                                                       │   old.txt    │                new.txt                 │
                                                                       │    sec/op    │    sec/op      vs base                 │
DatastoreDriver/cockroachdb-overlap-static/TestTuple/SnapshotRead-10     866.3µ ± 12%   1720.0µ ±  9%   +98.56% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-static/TestTuple/Touch-10            11.93m ±  7%    11.56m ±  6%         ~ (p=0.165 n=10)
DatastoreDriver/cockroachdb-overlap-static/TestTuple/Create-10           13.10m ± 12%    12.89m ±  5%         ~ (p=0.353 n=10)
DatastoreDriver/cockroachdb-overlap-insecure/TestTuple/SnapshotRead-10   876.6µ ±  9%   1807.8µ ±  7%  +106.23% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-insecure/TestTuple/Touch-10          4.223m ± 10%    5.215m ±  9%   +23.49% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-insecure/TestTuple/Create-10         3.834m ±  9%    4.699m ± 15%   +22.56% (p=0.000 n=10)
geomean                                                                  3.526m          4.741m         +34.47%

                                                                       │   old.txt    │               new.txt                │
                                                                       │     B/op     │     B/op      vs base                │
DatastoreDriver/cockroachdb-overlap-static/TestTuple/SnapshotRead-10     12.96Ki ± 0%   14.29Ki ± 0%  +10.30% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-static/TestTuple/Touch-10            16.51Ki ± 2%   17.48Ki ± 0%   +5.87% (p=0.001 n=10)
DatastoreDriver/cockroachdb-overlap-static/TestTuple/Create-10           14.51Ki ± 3%   15.85Ki ± 2%   +9.21% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-insecure/TestTuple/SnapshotRead-10   12.96Ki ± 0%   14.29Ki ± 0%  +10.29% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-insecure/TestTuple/Touch-10          14.52Ki ± 1%   15.95Ki ± 1%   +9.88% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-insecure/TestTuple/Create-10         12.93Ki ± 1%   14.29Ki ± 1%  +10.55% (p=0.000 n=10)
geomean                                                                  14.01Ki        15.31Ki        +9.34%

                                                                       │  old.txt   │              new.txt              │
                                                                       │ allocs/op  │ allocs/op   vs base               │
DatastoreDriver/cockroachdb-overlap-static/TestTuple/SnapshotRead-10     230.0 ± 0%   246.0 ± 0%  +6.96% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-static/TestTuple/Touch-10            237.0 ± 0%   253.0 ± 0%  +6.75% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-static/TestTuple/Create-10           232.0 ± 0%   248.0 ± 0%  +6.90% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-insecure/TestTuple/SnapshotRead-10   230.0 ± 0%   246.0 ± 0%  +6.96% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-insecure/TestTuple/Touch-10          221.0 ± 0%   237.0 ± 0%  +7.24% (p=0.000 n=10)
DatastoreDriver/cockroachdb-overlap-insecure/TestTuple/Create-10         216.0 ± 0%   232.0 ± 0%  +7.41% (p=0.000 n=10)
geomean
```

I have [another branch](https://github.com/authzed/spicedb/pull/1284) exploring ways to remove the overhead, but so far haven't been able to show it survive node drains completely in the way that this PR does.

Note that this requires admin permission on the cockroach database.