@ecordell 2023-05-06
Closing in favor of #1294

This approach was effective, but requires admin permission. We may revisit it as an option in the future, but for now it's better to just have one way to handle node drains.